Move all debug overlay code to its own module
Instead of scattering this throughout the codebase, all the dev overlay stuff is in one spot now. This makes it less likely we accidentally include any of this in a release build
